Refactor: NoteDetail 리팩토링
